requirements.txt
diff-match-patch==20121119
Django==1.11.3
django-import-export==0.5.1
django-ipware==1.1.6
django-jquery==3.1.0
django-wysiwyg-redactor==0.5.1
et-xmlfile==1.0.1
jdcal==1.3
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.9
Pillow==4.2.1
pytz==2017.2
PyYAML==3.12
selenium==3.4.3
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
